+++ src/main/java/com/example/bookstore_backend/dto/CartBookDto.java
@@ -5,15 +5,16 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class CartBookDto {
-    private Integer cid;
-    private Integer bid;
-    private Integer uid;
+    private Long cid;
+    private Long bid;
+    private Long uid;
     private String name;
     private Double price;
     private String pic;
