+++ src/main/java/com/example/bookstore_backend/service/CartService.java
@@ -6,8 +6,8 @@ import com.example.bookstore_backend.model.CartBook;
 import java.util.List;
 
 public interface CartService {
-    public List<CartBookDto> getCartItemByUid(Integer uid);
+    public List<CartBookDto> getCartItemByUid(Long uid);
     public Boolean AddCartItem(CartBook book);
     public Boolean updateCartItem(CartBook book);
-    public Boolean deleteCartItem(Integer cid);
+    public Boolean deleteCartItem(Long cid);
 }
